pill rename
