--- packages/ui-toolkit/src/components/atoms/Chip/index.ts
@@ -1,1 +1,0 @@
-export { default as Chip, IconPosition } from './Chip';
