+++ packages/ui-toolkit/src/components/atoms/Pill/index.ts
@@ -1,0 +1,1 @@
+export { default as Pill } from './Pill';
