+++ packages/ui-toolkit/src/components/atoms/index.ts
@@ -1,7 +1,7 @@
 export * from './Accordion';
 export * from './Button';
 export * from './CheckBox';
-export * from './Chip';
+export * from './Pill';
 export * from './DateSelector';
 export * from './Dropdown';
 export * from './GoBack';
